Skip if on compute nodes

+++ tmux_launch.sh
@@ -19,29 +19,34 @@
 
 
 # http://forums.gentoo.org/viewtopic-p-6351393.html#6351393
-# Make sure we are not already in a tmux session
-if [[ -z "$TMUX" ]]; then
-    timeout=2
-    echo "Press any key to cancel tmux launch (${timeout} seconds timeout)."
-    if read -s -n 1 -t ${timeout}; then
-        echo "tmux launch/attach canceled!"
-    else
-        tmux_nb=`tmux ls | wc -l`
-        if [[ "$tmux_nb" == "0" ]]; then
-            #echo "Launching tmux..."
-            tmux
+
+# Skip if on a compute nodes of unicron
+hostname=`hostname`
+if [[ "${hostname:0:4}" != "node" ]]; then
+    # Make sure we are not already in a tmux session
+    if [[ -z "$TMUX" ]]; then
+        timeout=2
+        echo "Press any key to cancel tmux launch (${timeout} seconds timeout)."
+        if read -s -n 1 -t ${timeout}; then
+            echo "tmux launch/attach canceled!"
         else
-            #echo "tmux already started, attaching..."
-            # Session is is date and time to prevent conflict
-            session_id=`date +%Y%m%d%H%M%S`
-            # Create a new session (without attaching it) and link to session id 0 (to share windows)
-            tmux new-session -d -t 0 -s $session_id
-            # Create a new window in that session
-            tmux new-window
-            # Attach to the new session
-            tmux attach-session -t $session_id
-            # When we detach from it, kill the session
-            tmux kill-session -t $session_id
+            tmux_nb=`tmux ls | wc -l`
+            if [[ "$tmux_nb" == "0" ]]; then
+                #echo "Launching tmux..."
+                tmux
+            else
+                #echo "tmux already started, attaching..."
+                # Session is is date and time to prevent conflict
+                session_id=`date +%Y%m%d%H%M%S`
+                # Create a new session (without attaching it) and link to session id 0 (to share windows)
+                tmux new-session -d -t 0 -s $session_id
+                # Create a new window in that session
+                tmux new-window
+                # Attach to the new session
+                tmux attach-session -t $session_id
+                # When we detach from it, kill the session
+                tmux kill-session -t $session_id
+            fi
         fi
     fi
 fi
